+++ src/editor/examples/lottieExample.h
@@ -109,7 +109,7 @@ struct LottieExample : tvgexam::Example
 
 		// Background
 		auto shape = tvg::Shape::gen();
-		shape->appendRect(0, 0, w, h);
+		shape->appendRect(0, 0, (float) w, (float) h);
 		shape->fill(75, 75, 75);
 		canvas->push(shape);
